% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{assessment}
\alias{assessment}
\title{Assessment information}
\format{
An object of class \code{data.frame} with 3 rows and 10 columns.
}
\usage{
assessment
}
\description{
A data set containing imaginary assessment information for three example species A, B and C. Data structure is
set out in such way, to simulate the data standards of the IUCN Red List of Threatened Species.
Any similarity to real data from the IUCN Red List of Threatened Species is a mere coincidence.
The variables (find more details for the variables in the IUCN Red List of Threatened Species website
https://www.iucnredlist.org/) are as follows:
}
\details{
@format Object of class sf and data frame with 3 rows and 10 variables:\code{assessmentId},
\code{internalTaxonId}, \code{scientificName}, \code{redlistCategory}, \code{redlistCriteria},
\code{yearPublished}, \code{assessmentDate}, \code{criteriaVersion}, \code{language},
\code{systems}
\describe{
\item{assessmentId}{unique id of the assessment of a species}
\item{internalTaxonId}{unique id of a species}
\item{scientificName}{scientific name of a species}
\item{redlistCategory}{}
\item{redlistCriteria}{}
\item{yearPublished}{year of the assessment publish on the IUCN Red List website}
\item{assessmentDate}{}
\item{criteriaVersion}{}
\item{language}{}
\item{systems}{}
}
source Generated in EXCEL
}
\keyword{datasets}
