% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{common_names}
\alias{common_names}
\title{Common names of species}
\format{
An object of class \code{data.frame} with 5 rows and 5 columns.
}
\usage{
common_names
}
\description{
A data set containing imaginary common names for three example species A, B and C. Data structure is
set out in such way, to simulate the data standards of the IUCN Red List of Threatened Species.
Any similarity to real data from the IUCN Red List of Threatened Species is a mere coincidence.
The variables (find more details for the variables in the IUCN Red List of Threatened Species website
https://www.iucnredlist.org/) are as follows:
}
\details{
@format Object of class sf and data frame with 5 rows and 5 variables:\code{internalTaxonId},
\code{scientificName}, \code{name}, \code{language}, \code{main}
\describe{
\item{internalTaxonId}{unique id of a species}
\item{scientificName}{scientific name of a species}
\item{name}{common name}
\item{language}{}
\item{main}{}
}
source Generated in EXCEL
}
\keyword{datasets}
