% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_repo.R
\name{create_repo}
\alias{create_repo}
\title{Creates a specific folder structure to perform KBA scoping analysis.}
\usage{
create_repo(x)
}
\arguments{
\item{x}{A directory under which inputs and outputs of a KBA scoping
analysis will be stored.}
}
\value{
Folder structure at a selected directory is created.
}
\description{
Creates a specific folder structure to perform KBA scoping analysis.
}
\examples{
x<- "C:/projects/KBA_scoping_analysis"

}
