% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/criteria.R
\name{criteria}
\alias{criteria}
\title{Combines criteria columns.}
\usage{
criteria(
  x,
  KBAcriteria = c("Criterion_A1", "Criterion_B1", "Criterion_B2", "Criterion_B3",
    "Criterion_D1", "Criterion_D2", "Criterion_D3")
)
}
\arguments{
\item{x}{Object of class sf and data frame representing species data.}

\item{KBAcriteria}{Object of class character to identify the KBAcriteria t
ested against the data.Takes values "Criterion_A1",
"Criterion_B1","Criterion_B2","Criterion_B3",
"Criterion_D1","Criterion_D2" and "Criterion_D3".}
}
\value{
x with one new column for all the KBA criteria
}
\description{
Combines criteria columns.
}
\examples{
df<- data.frame(ScientificName=c("A"),Criterion_A1=c("A1a"),Criterion_B1=c("B1")) \%>\%
  cbind(KBAscope::species[1,-3])

criteria(df, KBAcriteria=c("Criterion_A1","Criterion_B1"))


}
