% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/criterion_B.R
\name{criterion_B}
\alias{criterion_B}
\title{Tests against KBA criteria for B:Geographically Restricted Biodiversity}
\usage{
criterion_B(x, criterion = c("B1", "B2", "B3"))
}
\arguments{
\item{x}{Object of class sf and data frame with all relevant information to
apply KBA criteria B1, B2 and B3.}

\item{criterion}{Object of class character. Takes values "B1","B2" and/or"B3".}
}
\value{
x with three new columns Criterion_B1, Criterion_B2, Criterion_B3.
}
\description{
Tests against KBA criteria for B:Geographically Restricted Biodiversity
}
\examples{
df<- data.frame(SCI_NAME= "A",Range_Restricted="", Eco_BioRestricted="",
  TaxonomicGroup="", proportion="") \%>\% dplyr::left_join(KBAscope::species)

criterion_B(df)

}
