% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/criterion_D.R
\name{criterion_D}
\alias{criterion_D}
\title{ests against KBA criteria for D: Biological Processes}
\usage{
criterion_D(x)
}
\arguments{
\item{x}{Object of class sf and data frame with all relevant information to
apply KBA criterion A.}
}
\value{
x with three new columns Criterion_D1, Criterion_D2 and Criterion_D3.
}
\description{
ests against KBA criteria for D: Biological Processes
}
\examples{
x<- data.frame(SCI_NAME= "A", SiteID="1001",proportion="1.3") \%>\%
            dplyr::left_join(KBAscope::species)

  criterion_D(x)



}
