% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_form_multi_site}
\alias{data_form_multi_site}
\title{Multi-site KBA data form}
\format{
An object of class \code{Workbook} of length 1.
}
\usage{
data_form_multi_site
}
\description{
KBA Single/multi-site Proposal form 1.2 that can be used to test data acquired
for a site to ensure it meets the KBA criteria as well as help a proposer compile
all the necessary documentation. The variables are as follows:
}
\details{
@format Object of class Workbook:
source Downloaded from the WDKBA website
(https://www.keybiodiversityareas.org/working-with-kbas/proposing-updating/criteria-tools)
}
\keyword{datasets}
