% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eco_bio_check.R
\name{eco_bio_check}
\alias{eco_bio_check}
\title{Identifies Whether a Species is Ecoregion or Bioregion Restricted}
\usage{
eco_bio_check(
  x,
  ecoregion = NULL,
  eco_name = NULL,
  bioregion = NULL,
  bio_name = NULL
)
}
\arguments{
\item{x}{Object of class sf and data frame representing species data.}

\item{ecoregion}{Object of class sf and data frame that represents relevant ecoregion.}

\item{eco_name}{Object of class character. The column that represents the
ecoregion names.}

\item{bioregion}{Object of class sf and data frame that represents relevant bioregion.}

\item{bio_name}{Object of class character. The column that represents the
bioregion names.}
}
\value{
object x with two new columns "Eco_BioRestricted" and "Eco_BioName"
}
\description{
Identifies Whether a Species is Ecoregion or Bioregion Restricted
}
\examples{

x<- KBAscope::species[1,] \%>\% dplyr::left_join(data.frame(ScientificName= "A",
  SCI_NAME="A", Range_Restricted="", Eco_BioRestricted="", TaxonomicGroup="",
  proportion=""))

eco_info<- data.frame(scientific_name= "A",Restricted.to.Terrestrial.ecoregion="TRUE",
  Restricted.to.a.FW.ecoregion="",Restricted.to.marine.ecoregion="",
  Restricted.to.FW.bioregion="",Restricted.to.Marine.Bioregion="")

eco_bio_check(x,ecoregion=KBAscope::ecoregion,eco_name="ECO_NAME",
  bioregion=NULL, bio_name=NULL)



}
