% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{eco_bioregion_restricted}
\alias{eco_bioregion_restricted}
\title{Ecoregion Bioregion restricted}
\format{
An object of class \code{data.frame} with 18901 rows and 24 columns.
}
\usage{
eco_bioregion_restricted
}
\description{
List of Ecoregion/bioregion restricted species: Criterion B3 requires users
to identify ecoregion-restricted or bioregion-restricted species at a site.
Here is a list of species that meet eco/bioregion restricted status for
terrestrial, freshwater and marine eco/bioregions. The variables are as follows:
}
\details{
@format Object of class data frame with 18901 rows and 24 variables:\code{taxid},
\code{scientific_name}, \code{group_level}, \code{Taxonomic.group},
\code{kingdom}, \code{phylum},\code{class},\code{order},
\code{family},\code{genus},\code{terrestrial.species},
\code{Freshwater.species},\code{Marine.species},
\code{B3a.or.B3b},\code{Restricted.to.Terrestrial.ecoregion},
\code{terrestrial_ecoregion_name},\code{Restricted.to.a.FW.ecoregion},
\code{freshwater_ecoregion_name},\code{Restricted.to.marine.ecoregion},
\code{marine_ecoregion_name},\code{Restricted.to.FW.bioregion},
\code{freshwater_bioregion_name},\code{Restricted.to.Marine.Bioregion},
\code{marine_bioregion_name}
\describe{
\item{taxid}{}
\item{scientific_name}{}
\item{group_level}{}
\item{Taxonomic.group}{}
\item{kingdom}{}
\item{phylum}{}
\item{class}{}
\item{order}{}
\item{family}{}
\item{genus}{}
\item{terrestrial.species}{}
\item{Freshwater.species}{}
\item{Marine.species}{}
\item{B3a.or.B3b}{}
\item{Restricted.to.Terrestrial.ecoregion}{}
\item{terrestrial_ecoregion_name}{}
\item{Restricted.to.a.FW.ecoregion}{}
\item{freshwater_ecoregion_name}{}
\item{Restricted.to.marine.ecoregion}{}
\item{marine_ecoregion_name}{}
\item{Restricted.to.FW.bioregion}{}
\item{freshwater_bioregion_name}{}
\item{Restricted.to.Marine.Bioregion}{}
\item{marine_bioregion_name}{}
}
source Downloaded from the WDKBA website
(https://www.keybiodiversityareas.org/working-with-kbas/proposing-updating/criteria-tools)
}
\keyword{datasets}
