% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ecoregion}
\alias{ecoregion}
\title{Ecoregion polygon}
\format{
An object of class \code{sf} (inherits from \code{data.frame}) with 1 rows and 3 columns.
}
\usage{
ecoregion
}
\description{
A data set containing imaginary ecoregion polygon. Created for testing.
}
\details{
@format  Object of class sf and data frame with 1 rows and 3 variables:\code{ECO_NAME},
\code{ID}, \code{geom}
\describe{
\item{ECO_NAME}{name of the ecoregion}
\item{ID}{unique id}
\item{geom}{geometry of the polygons}
}
source Generated in QGIS 3.28.2
}
\keyword{datasets}
