% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/potential_kbas.R
\name{potential_kbas}
\alias{potential_kbas}
\title{Generates potential KBA sites.}
\usage{
potential_kbas(x, system = "terrestrial", output = ".gpkg")
}
\arguments{
\item{x}{The directory under which inputs and outputs of a KBA scoping
analysis are stored.}

\item{system}{Object of class character. Takes values "terrestrial",
"freshwater" or "marine".}

\item{output}{Object of class character. Defines the output file type.
Default is ".gpkg" which stands for GEOPACKAGE.}
}
\value{
Creates output files 'potential_KBAs.gpkg', 'ptrigger_species.gpkg'
and  'ptrigger_species.csv' under results folder.
}
\description{
Generates potential KBA sites.
}
\examples{
if (FALSE) {

potential_kbas(x)

}

}
