% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/red_list_info.R
\name{red_list_info}
\alias{red_list_info}
\title{Combines the required information from the IUCN Red List of Threatened Species}
\usage{
red_list_info(taxonomy, assessment, common_names)
}
\arguments{
\item{taxonomy}{Object of class data frame from the IUCN Red List website
that contains the taxonomy of species.}

\item{assessment}{Object of class data frame from the IUCN Red List website
that contains the details of the species assessments.}

\item{common_names}{Object of class data frame from the IUCN Red List website
that contains the common names of species.}
}
\value{
Object of class data frame that includes the information from taxonomy,
assessments and common names files combined.
}
\description{
Combines the required information from the IUCN Red List of Threatened Species
}
\examples{
red_list_info(KBAscope::taxonomy, KBAscope::assessment, KBAscope::common_names)



}
