% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results_summary.R
\name{results_summary}
\alias{results_summary}
\title{Generates summary file of results.}
\usage{
results_summary(x, path = NULL)
}
\arguments{
\item{x}{Object of class data frame. The "trigger_species.csv" file
generated from KBAscope::potential_kbas.}

\item{path}{Path of the directory to save outputs.}
}
\value{
Creates output "results_summary.xlsx" under results folder.
}
\description{
Generates summary file of results.
}
\examples{

\dontrun{
  df<- data.frame(Criteria="",SiteID="",TaxonomicGroup="",ScientificName="",
    Criterion_A1="",Criterion_B1="",Criterion_B2="",Criterion_B3="",
    site_B2="",B2_RR="",nB2="",site_B3="",nB3="")

  results_summary(df, path="C:/projects/KBA_scoping_analysis/results/")
}   


}
