% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_triggers.R
\name{save_triggers}
\alias{save_triggers}
\title{Saves output per species triggering KBA criteria}
\usage{
save_triggers(
  x,
  type = c("AOO", "AOH", "localities", "range_maps", "population"),
  path = NULL,
  output = ".gpkg"
)
}
\arguments{
\item{x}{Object of class sf and data frame representing species data.}

\item{type}{Object of class character to identify the type of distribution
data of x.Takes values "AOO", "AOH","localities",
"population" or "range_maps".}

\item{path}{Path of the directory to save outputs.}

\item{output}{Object of class character. Defines the output file type.
Default is ".gpkg" which stands for GEOPACKAGE.}
}
\value{
Saves outputs of potential KBA sites per species that triggers
KBA criteria.
}
\description{
Saves output per species triggering KBA criteria
}
\examples{
if(FALSE){
x<- KBAscope::species[1,-3] \%>\% cbind(data.frame(ScientificName=c("A"),
          Criteria=c("A1a,B1")))

save_triggers(x, type="range_maps",path=NULL, output=".gpkg")

}

}
