% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/site_calculation.R
\name{site_calculation}
\alias{site_calculation}
\title{Calculates the site population size.}
\usage{
site_calculation(
  x,
  spatial_units,
  type = c("AOO", "AOH", "localities", "range_maps", "population")
)
}
\arguments{
\item{x}{Object of class sf and data frame representing species data.}

\item{spatial_units}{Object of class sf and data frame that represent the
spatial unites data.}

\item{type}{Object of class character to identify the type of distribution
data of x.Takes values "AOO", "AOH","localities",
"population" or "range_maps".}
}
\value{
The intersection of x and spatial_units.
}
\description{
Calculates the site population size.
}
\examples{

spatial_units<- KBAscope::spatial_units \%>\% dplyr::rename(SiteID= id) \%>\%
  dplyr::mutate(SiteRange=10,GlobalRange=100)

site_calculation(KBAscope::species, spatial_units, type="range_maps")


}
