% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{spatial_units}
\alias{spatial_units}
\title{Spatial units polygons}
\format{
An object of class \code{sf} (inherits from \code{data.frame}) with 486 rows and 2 columns.
}
\usage{
spatial_units
}
\description{
A data set containing grid cells of 100 x 100 km2.
The variables are as follows:
@format Object of class sf and data frame with 486 rows and 2 variables: \code{id}, \code{geom}
\describe{
\item{id}{unique id of each grid cell}
\item{geom}{geometry of the polygons}
}
source Generated in QGIS 3.28.2
}
\keyword{datasets}
