% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/species_edit.R
\name{species_edit}
\alias{species_edit}
\title{Changes species data into format to apply KBA scoping analysis.}
\usage{
species_edit(
  x,
  type = c("AOO", "AOH", "localities", "range_maps"),
  name,
  species_info,
  system = "terrestrial"
)
}
\arguments{
\item{x}{Object of class sf and data frame representing species data.}

\item{type}{Object of class character to identify the type of distribution
data of x. Takes values "AOO", "AOH","localities" or "range_maps".}

\item{name}{Object of class character. The column that represents the
scientific name of species.}

\item{species_info}{Object of class data frame that results from the
function in KBAscope 'species_info'.}

\item{system}{Object of class character. Takes values "terrestrial",
"freshwater" or "marine".}
}
\value{
Object of class sf and data frame representing species data in
the format that a KBA scoping analysis can be performed.
}
\description{
Changes species data into format to apply KBA scoping analysis.
}
\examples{

if (FALSE){
x<- KBAscope::species \%>\% dplyr::mutate(ScientificName="A")
          RL<- red_list_info(KBAscope::taxonomy, KBAscope::assessment,
           KBAscope::common_names)

species_edit(x, type="range_maps", name="SCI_NAME",species_info=RL,
  system="terrestrial")
}


}
