% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/species_info.R
\name{species_info}
\alias{species_info}
\title{Creates a data frame with all the necessary information to apply KBA criteria.}
\usage{
species_info(
  x,
  name,
  type = c("AOO", "AOH", "localities", "range_maps"),
  red_list = TRUE,
  taxonomy_info = NULL
)
}
\arguments{
\item{x}{Object of class data frame that contains the list of species.
If the species are not assessed in the IUCN Red List of Threatened
Species, x should contain the taxonomy of the species.}

\item{name}{Object of class character. The column containing the species name.}

\item{type}{Object of class character to identify the type of distribution
data of x. Takes values "AOO", "AOH","localities" or "range_maps".}

\item{red_list}{If TRUE data are assessed in the IUCN Red List of Threatened
Species. If FALSE data are not assessed in the IUCN Red List
of Threatened Species.}

\item{taxonomy_info}{Object of class data frame created by
KBAscope::red_list_info() or manually for species not assessed in
the IUCN Red List of Threatened Species.}
}
\value{
Object of class data frame that contains all the necessary
information to apply KBA criteria.
}
\description{
Creates a data frame with all the necessary information to apply KBA criteria.
}
\examples{
\dontrun{
  species_info(KBAscope::species,name="SCI_NAME", red_list=TRUE,info= info)
}


}
