% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{taxonomic_groups}
\alias{taxonomic_groups}
\title{Taxonomic groups}
\format{
An object of class \code{data.frame} with 354 rows and 12 columns.
}
\usage{
taxonomic_groups
}
\description{
Taxonomic ranks above the species level that are appropriate for applying
KBA criterion B. The variables are as follows:
@format Object of class sf and data frame with 354 rows and 12 variables:\code{Sort.Order},
\code{Kingdom}, \code{Phylum}, \code{Class}, \code{Order},
\code{Superfamily},\code{Taxonomic.group.level},
\code{Number.RR.species.required.at.KBA.with.1..range.for.B2},
\code{Area.for.25..B2.RR.definition..km2....otherwise.use.10.000.km2},
\code{Median.range.size.for.B3.assessment.for.comprehensively.assessed.species},
\code{Apply.Ecoregion..B3a..or.bioregion..B3b..restriction},
\code{Notes}
\describe{
\item{Sort.Order}{}
\item{Kingdom}{}
\item{Phylum}{}
\item{Class}{}
\item{Order}{}
\item{Superfamily}{}
\item{Taxonomic.group.level}{}
\item{Number.RR.species.required.at.KBA.with.1..range.for.B2}{}
\item{Area.for.25..B2.RR.definition..km2....otherwise.use.10.000.km2}{}
\item{Median.range.size.for.B3.assessment.for.comprehensively.assessed.species}{}
\item{Apply.Ecoregion..B3a..or.bioregion..B3b..restriction}{}
\item{Notes}{}
}
source Downloaded from the WDKBA website
(https://www.keybiodiversityareas.org/working-with-kbas/proposing-updating/criteria-tools)
}
\keyword{datasets}
