% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{taxonomy}
\alias{taxonomy}
\title{Taxonomy information}
\format{
An object of class \code{data.frame} with 3 rows and 7 columns.
}
\usage{
taxonomy
}
\description{
A data set containing imaginary taxonomic information for three example species A, B and C. Data structure is
set out in such way, to simulate the data standards of the IUCN Red List of Threatened Species.
Any similarity to real data from the IUCN Red List of Threatened Species is a mere coincidence.
The variables (find more details for the variables in the IUCN Red List of Threatened Species website
https://www.iucnredlist.org/) are as follows:
}
\details{
@format Object of class sf and data frame with 3 rows and 7 variables:\code{internalTaxonId},
\code{scientificName}, \code{kingdomName}, \code{phylumName}, \code{className}, \code{orderName},
\code{familyName}
\describe{
\item{internalTaxonId}{unique id of a species}
\item{scientificName}{scientific name of a species}
\item{kingdomName}{}
\item{phylumName}{}
\item{className}{}
\item{orderName}{}
\item{familyName}{}
}
source Generated in EXCEL
}
\keyword{datasets}
