% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{triggers_gis}
\alias{triggers_gis}
\title{Triggers GIS}
\format{
Object of class sf and data frame with 92 rows and 39 variables:
\code{SiteID},
\code{LEGEND},
\code{ScientificName},
\code{internalTaxonId},
\code{CommonName},
\code{TaxonomicGroup_KBA_dataForm},
\code{GlobalRedListCategory},
\code{AssessAgainstA1c_A1d},
\code{AssessmentParameter},
\code{Source},
\code{SiteID},
\code{SiteID},
\code{SiteID},
\code{SiteID},
\code{SiteID},
\code{SiteID},
\code{SiteID},
\code{SiteID},
\code{SiteID},
\code{SiteID},
\code{SiteID},
\code{SiteID},
\code{SiteID},
\code{SiteID},
\code{geom}
\describe{
\item{id}{unique id of each grid cell}
\item{geom}{geometry of the polygons}
}
}
\usage{
triggers_gis
}
\description{
KBAscope test species that trigger KBA criteria in KBAscope test spatial units.
It is added to KBAscope to increase the understanding of applying KBAscope and #
to assist examples in several functions.
}
\keyword{datasets}
