% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view_results.R
\name{view_results}
\alias{view_results}
\title{Visualize results.}
\usage{
view_results(x, y)
}
\arguments{
\item{x}{Object of class data frame. The "ptrigger_species.csv" file
generated from KBAscope::potential_kbas.}

\item{y}{Object of class sf and data frame. The "ptrigger_species.gpkg"
file generated from KBAscope::potential_kbas.}
}
\value{
Creates figures of summary results.
}
\description{
Visualize results.
}
\examples{
if(FALSE){

x<- x
}

}
