% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kba_dataforms.R
\name{kba_dataforms}
\alias{kba_dataforms}
\title{Pre-populates KBA data forms with KBA scoping analysis results.}
\usage{
kba_dataforms(x)
}
\arguments{
\item{x}{Object of class data frame. The "trigger_species.csv" file generated
from KBAscope::potential_kbas.}
}
\value{
Generates pre-populated KBA data forms under results folder.
}
\description{
Pre-populates KBA data forms with KBA scoping analysis results.
}
\examples{
\dontrun{
  df<- sf::st_drop_geometry(triggers_gis)

  kba_dataforms(df)
}

}
