% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view_results.R
\name{view_results}
\alias{view_results}
\title{Visualize results.}
\usage{
view_results(x, y, pic = TRUE, pic_size = 6, font_size = 9)
}
\arguments{
\item{x}{Object of class data frame. The "ptrigger_species.csv" file
generated from KBAscope::potential_kbas.}

\item{y}{Object of class sf and data frame. The "ptrigger_species.gpkg"
file generated from KBAscope::potential_kbas.}

\item{pic}{Object of class logical. Takes values TRUE for adding a taxonomy picture or FALSE for not adding a taxonomy picture.}

\item{pic_size}{Object of class numeric that defines the size of taxonomy picture.}

\item{font_size}{Object of class numeric that defines the font size of x axis text.}
}
\value{
Creates figures of summary results.
}
\description{
Visualize results.
}
\examples{
\dontrun{
  df<- sf::st_drop_geometry(triggers_gis)
  view_results(df,triggers_gis, pic=TRUE)
}

}
